refactor(desktop): extract session-overview subcomponents to files
move 8 inline subcomponents from the overview pane to sibling files

(970 to 604 lines); pure mechanical, zero behavior change.
