+++ apps/desktop/src/features/session/components/SessionOverviewPane/StartRowContent.tsx
@@ -1,0 +1,39 @@
+import { ArrowRight } from 'lucide-react';
+import type { LucideIcon } from 'lucide-react';
+import { Chip, cn, tintClasses } from '@goodboy/ui';
+import type { Tone } from '@goodboy/ui';
+
+type Props = {
+  readonly icon: LucideIcon;
+  readonly tone: Tone;
+  readonly label: string;
+  readonly description: string;
+  readonly chip?: boolean;
+};
+
+export const StartRowContent = ({ icon: Icon, tone, label, description, chip }: Props) => (
+  <>
+    <span
+      aria-hidden
+      className={cn(
+        'flex size-9 shrink-0 items-center justify-center rounded-lg ring-1',
+        tintClasses(tone).bg,
+        tintClasses(tone).ring,
+      )}
+    >
+      <Icon size={16} aria-hidden className={tintClasses(tone).icon} />
+    </span>
+    <span className="flex min-w-0 flex-1 flex-col gap-0.5">
+      <span className="flex items-center gap-1.5 text-sm font-medium text-foreground">
+        {label}
+        {chip ? <Chip tone="accent" size="sm" label="recommended" /> : null}
+      </span>
+      <span className="truncate text-2xs text-muted-foreground">{description}</span>
+    </span>
+    <ArrowRight
+      size={15}
+      aria-hidden
+      className="shrink-0 text-muted-foreground/30 motion-safe:transition-transform group-hover:translate-x-0.5 group-hover:text-muted-foreground"
+    />
+  </>
+);
